WEBAPP-316: separate matcher from react component, remove html-to-text, add some more tests
