+++ package.json
@@ -39,7 +39,6 @@
     "focus-trap-react": "^6.0.0",
     "hashids": "^2.2.1",
     "history": "^4.9.0",
-    "htmlparser2": "^3.9.2",
     "i18next": "^19.1.0",
     "i18next-browser-languagedetector": "^4.0.1",
     "lodash": "^4.17.19",
@@ -117,8 +116,8 @@
     "flow-bin": "~0.117.0",
     "flow-typed": "^2.5.1",
     "html-loader": "^0.5.5",
-    "html-to-text": "^5.1.1",
     "html-webpack-plugin": "^3.2.0",
+    "htmlparser2": "^3.9.2",
     "identity-obj-proxy": "^3.0.0",
     "img-loader": "^3.0.1",
     "jest": "^24.9.0",
