+++ src/routes/categories/components/__tests__/CategoryEntry.spec.js
@@ -33,18 +33,6 @@ const childCategory = new CategoryModel({
   lastUpdate: moment('2017-11-18T19:30:00.000Z'),
   hash: 'a36a57'
 })
-const noContentCategory = new CategoryModel({
-  root: false,
-  path: '/augsburg/de/test',
-  title: 'Child',
-  content: '',
-  parentPath: '/augsburg/de',
-  order: 11,
-  availableLanguages: new Map([['en', '4861'], ['ar', '4867'], ['fa', '4868']]),
-  thumbnail: 'https://cms.integreat-ap…03/Beratung-150x150.png',
-  lastUpdate: moment('2017-11-18T19:30:00.000Z'),
-  hash: 'a36a57'
-})
 const noThumbCategory = new CategoryModel({
   root: false,
   path: '/augsburg/de/willkommen/willkommen-in-augsburg',
@@ -100,119 +88,8 @@ describe('CategoryEntry', () => {
     expect(findComponent(wrapper, 'CategoryThumbnail').prop('src')).toEqual(iconPlaceholder)
   })
 
-  describe('getWords', () => {
-    const wrapper = shallow(
-      <CategoryEntry theme={brightTheme} category={noThumbCategory} subCategories={[noThumbCategory]} />
-    ).dive()
-
-    it('should return empty list for an empty input', () => {
-      const testString = ''
-      const categoryEntry = wrapper.instance()
-      expect(categoryEntry.getWords(testString)).toEqual([])
-    })
-
-    it('should split text into words and remove all whitespace characters', () => {
-      const testString = ' a \tb \r\nc\n\nd-e ,f :g /h'
-      const categoryEntry = wrapper.instance()
-      expect(categoryEntry.getWords(testString)).toEqual(['a', 'b', 'c', 'd-e', ',f', ':g', '/h'])
-    })
-  })
-
   describe('getMatchedContent', () => {
-    const numWords = 3
-
-    describe('getContentBeforeMatchIndex', () => {
-      it('should return 3 words before the specified index an start of the word', () => {
-        const content = 'This is some test content'
-        const matchIdx = 15
-        const wrapper = shallow(
-            <CategoryEntry theme={brightTheme}
-                           category={category}
-                           contentWithoutHtml={category.content}
-                           subCategories={[]} />
-        ).dive()
-        const categoryEntry = wrapper.instance()
-        expect(categoryEntry.getContentBeforeMatchIndex(content, matchIdx, false, numWords))
-          .toBe('This is some te')
-      })
-
-      it('should return 3 words before the specified index', () => {
-        const content = 'This is some test content'
-        const matchIdx = 13
-        const wrapper = shallow(
-          <CategoryEntry theme={brightTheme}
-                         category={category}
-                         contentWithoutHtml={category.content}
-                         subCategories={[]} />
-        ).dive()
-        const categoryEntry = wrapper.instance()
-        expect(categoryEntry.getContentBeforeMatchIndex(content, matchIdx, true, numWords))
-          .toBe('This is some ')
-      })
-    })
-
-    describe('getMatchedContentAfterMatchIndex', () => {
-      it('should return 3 words before the specified index an start of the word', () => {
-        const content = 'This is some test content'
-        const matchIdx = 1
-        const wrapper = shallow(
-          <CategoryEntry theme={brightTheme}
-                         category={category}
-                         contentWithoutHtml={category.content}
-                         subCategories={[]} />
-        ).dive()
-        const categoryEntry = wrapper.instance()
-        expect(categoryEntry.getContentAfterMatchIndex(content, matchIdx, numWords))
-          .toBe('his is some test')
-      })
-
-      it('should return 3 words before the specified index', () => {
-        const content = 'This is some test content'
-        const matchIdx = 0
-        const wrapper = shallow(
-          <CategoryEntry theme={brightTheme}
-                         category={category}
-                         contentWithoutHtml={category.content}
-                         subCategories={[]} />
-        ).dive()
-        const categoryEntry = wrapper.instance()
-        expect(categoryEntry.getContentAfterMatchIndex(content, matchIdx, numWords))
-          .toBe('This is some test')
-      })
-    })
-
-    it('should return null for undefined query', () => {
-      const wrapper = shallow(
-        <CategoryEntry theme={brightTheme}
-                       category={noThumbCategory}
-                       contentWithoutHptml={noThumbCategory.content}
-                       subCategories={[]} />
-      ).dive()
-      const categoryEntry = wrapper.instance()
-      expect(categoryEntry.getMatchedContent(numWords)).toBeNull()
-    })
-
-    it('should return null for categories without content', () => {
-      const wrapper = shallow(
-        <CategoryEntry theme={brightTheme} category={noContentCategory} query='abc' subCategories={[]} />
-      ).dive()
-      const categoryEntry = wrapper.instance()
-      expect(categoryEntry.getMatchedContent(numWords)).toBeNull()
-    })
-
-    it('should return null for empty query', () => {
-      const wrapper = shallow(
-        <CategoryEntry theme={brightTheme}
-                       category={noThumbCategory}
-                       contentWithoutHtml={noThumbCategory.content}
-                       query=''
-                       subCategories={[]} />
-      ).dive()
-      const categoryEntry = wrapper.instance()
-      expect(categoryEntry.getMatchedContent(numWords)).toBeNull()
-    })
-
-    it('should return the match with query starting at the beginning of a word', () => {
+    it('should return the highlighted match item', () => {
       const query = 'test'
       const selectedSection = 'this is a test content which is'
       const numWords = 3
@@ -233,34 +110,20 @@ describe('CategoryEntry', () => {
       expect(contentMatchProps.textToHighlight).toEqual(selectedSection)
     })
 
-    it('should match the same for query matching whole word and query matching in between a word', () => {
-      const completeWordQuery = 'test'
-      const inBetweenWordQuery = 'es'
+    it('should not return a match item', () => {
+      const query = 'test'
+      const numWords = 3
       const wrapper = shallow(
         <CategoryEntry theme={brightTheme}
                        category={category}
                        contentWithoutHtml={category.content}
-                       query={completeWordQuery}
+                       query={query}
                        subCategories={[]} />
       ).dive()
       const categoryEntry = wrapper.instance()
-      const contentMatchItem = shallow(categoryEntry.getMatchedContent(numWords))
-      const contentMatchProps = contentMatchItem.props()
-
-      const wrapperWithDifferentQuery = shallow(
-        <CategoryEntry theme={brightTheme}
-                       category={category}
-                       contentWithoutHtml={category.content}
-                       query={inBetweenWordQuery}
-                       subCategories={[]} />
-      ).dive()
-      const categoryEntryWithDifferentQuery = wrapperWithDifferentQuery.instance()
-      const contentMatchItemWithDifferentQuery = shallow(categoryEntryWithDifferentQuery.getMatchedContent(numWords))
-      const contentMatchPropsWithDifferentQuery = contentMatchItemWithDifferentQuery.props()
-
-      expect(contentMatchPropsWithDifferentQuery.searchWords).toHaveLength(1)
-      expect(contentMatchPropsWithDifferentQuery.searchWords).toContain(inBetweenWordQuery)
-      expect(contentMatchPropsWithDifferentQuery).toEqual({ ...contentMatchProps, searchWords: ['es'] })
+      categoryEntry.contentMatcher.getMatchedContent = jest.fn()
+      categoryEntry.contentMatcher.getMatchedContent.mockReturnValue(null)
+      expect(categoryEntry.getMatchedContent(numWords)).toBeNull()
     })
   })
 })
